Safari-609.1.11.1/LayoutTests/fast/files/not-enough-arguments-expected.txt

This test tries calling various FileAPI functions with too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS reader.readAsArrayBuffer() threw exception TypeError: Not enough arguments.
PASS reader.readAsBinaryString() threw exception TypeError: Not enough arguments.
PASS reader.readAsText() threw exception TypeError: Not enough arguments.
PASS reader.readAsDataURL() threw exception TypeError: Not enough arguments.
PASS fileInput.files.item() threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
